refactor: make consistent adapter names for version detection
