+++ mapping_suite_sdk/mapping_package_v1/adapters/mp_v1_version_detector.py
