+++ mapping_suite_sdk/mapping_package_v2/adapters/mp_v2_version_detector.py
