+++ mapping_suite_sdk/mapping_package_v3/adapters/mp_v3_version_detector.py
